requirements.txt
dependency_injector==4.41.0
discord.py==2.5.2
oauth2client==4.1.3
python-dotenv==1.0.1
requests==2.28.2
